split some commands
