+++ vs/mob.vcxproj
@@ -58,7 +58,13 @@
     <Link />
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\src\cmd\build.cpp" />
+    <ClCompile Include="..\src\cmd\cmake.cpp" />
     <ClCompile Include="..\src\cmd\commands.cpp" />
+    <ClCompile Include="..\src\cmd\git.cpp" />
+    <ClCompile Include="..\src\cmd\list.cpp" />
+    <ClCompile Include="..\src\cmd\release.cpp" />
+    <ClCompile Include="..\src\cmd\tx.cpp" />
     <ClCompile Include="..\src\core\conf.cpp" />
     <ClCompile Include="..\src\core\context.cpp" />
     <ClCompile Include="..\src\core\env.cpp" />
